requirements.txt
streamlit
tavily-python
google-genai
weasyprint
markdown2